+++ lib/ReactViews/Story/StoryEditor.jsx
@@ -6,7 +6,7 @@ import { withTranslation } from "react-i18next";
 import tinymce from "tinymce";
 
 // Lazy load the Editor component as the tinyMCE library is large
-const Editor = React.lazy(() => import("../Generic/Editor.jsx"));
+const Editor = React.lazy(() => import("../Generic/Editor.tsx"));
 class StoryEditor extends React.Component {
   constructor(props) {
     super(props);
@@ -186,6 +186,7 @@ class StoryEditor extends React.Component {
           <div className={Styles.body}>
             <Suspense fallback={<div>Loading...</div>}>
               <Editor
+                language={this.props.i18n.language}
                 html={this.state.text}
                 onChange={(_newValue, editor) => {
                   // TODO: This makes StoryEditor tightly coupled to Editor. How to reduce coupling?
@@ -196,7 +197,6 @@ class StoryEditor extends React.Component {
                   const text = editor.getBody().innerHTML;
                   this.setState({ text });
                 }}
-                terria={this.props.terria}
               />
             </Suspense>
           </div>
@@ -212,6 +212,7 @@ StoryEditor.propTypes = {
   saveStory: PropTypes.func,
   exitEditingMode: PropTypes.func,
   t: PropTypes.func.isRequired,
+  i18n: PropTypes.object,
   terria: PropTypes.object
 };