+++ lib/ThirdParty/global.d.ts
@@ -4,3 +4,4 @@ declare module "*.xml";
 declare module "*.svg";
 declare module "*.png";
 declare module "*.jpg";
+declare module "*.css";